component_b/pyproject.toml
[project]
name = "distributed-transcription-system"
version = "0.1.0"
description = "LLM-agnostic distributed transcription processing system"
authors = [
    {name = "Pablo Toledo", email = "pablo@example.com"}
]
requires-python = ">=3.10"
dependencies = [
    "fast-agent-mcp",
    "sentence-transformers",
    "hdbscan",
    "scikit-learn",
    "transformers",
    "torch",
    "pydantic>=2.11.7",
    "nltk>=3.9.1",
    "python-pptx>=0.6.21",
    "PyPDF2>=3.0.0",
    "Pillow>=10.0.0",
    "tiktoken>=0.5.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100
